upgrade graph & create wallet 2
