+++ pubspec.yaml
@@ -1,16 +1,16 @@
 name: wallet_core
 author: Lior Rabin <lior9586@gmail.com>
-version: 0.0.1+66
+version: 0.0.1+67
 homepage: https://github.com/fuseio/wallet_core
 description: Fuse wallet core Dart library to interact with Ethereum based networks and APIs.
-environment: 
-  sdk: '>=2.0.0-dev.68.0 <3.0.0'
-dependencies: 
+environment:
+  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+dependencies:
   bip32: ^1.0.5
   bip39: ^1.0.3
   hex: ^0.1.2
-  http: ^0.12.0
+  http: ^0.12.0+4
   resource: ^2.1.6
-  web3dart: ^1.2.1
-  graphql: ^2.0.0
+  web3dart: ^1.2.3
+  graphql: ^3.0.0
   decimal: ^0.3.5
